Update README.md
Objective: Develop a custom Linux shell

The objective of this C++ program is to:

•Appreciate how the bash shell runs commands
•Gain familiarity with fork and exec system calls
•Continue to build strength with I/O streams & string processing
•Reviewproblem solving strategies.


Background:

Operating Systems essentially provides fork and exec system calls to run programs, but an OS 
does not start running programs by itself. This task is delegated to another program, called 
a "shell". The shell (in our case we are using bashshell) accepts inputs from the user and 
based on the user-input executes different commands. Note that shells can be graphical and permit 
users to double-click on an icon to indicate the program to run.